cripto/cryptocsanner/requirements.txt
pyTelegramBotAPI
ccxt
python-dotenv